+++ pytorch_lightning/trainer/trainer.py
@@ -29,6 +29,7 @@ from pytorch_lightning.profiler import BaseProfiler
 from pytorch_lightning.trainer.callback_hook import TrainerCallbackHookMixin
 from pytorch_lightning.trainer.configuration_validator import ConfigValidator
 from pytorch_lightning.trainer.data_loading import TrainerDataLoadingMixin
+from pytorch_lightning.trainer.deprecated_api import TrainerDeprecatedAPITillVer0_11
 from pytorch_lightning.trainer.logging import TrainerLoggingMixin
 from pytorch_lightning.trainer.model_hooks import TrainerModelHooksMixin
 from pytorch_lightning.trainer.optimizers import TrainerOptimizersMixin
@@ -78,6 +79,7 @@ class Trainer(
     TrainerLoggingMixin,
     TrainerTrainingTricksMixin,
     TrainerDataLoadingMixin,
+    TrainerDeprecatedAPITillVer0_11,
 ):
     def __init__(
         self,
@@ -108,8 +110,8 @@ class Trainer(
         limit_val_batches: Union[int, float] = 1.0,
         limit_test_batches: Union[int, float] = 1.0,
         val_check_interval: Union[int, float] = 1.0,
-        log_save_interval: int = 100,
-        row_log_interval: int = 50,
+        flush_logs_every_n_steps: int = 100,
+        log_every_n_steps: int = 50,
         distributed_backend: Optional[str] = None,
         sync_batchnorm: bool = False,
         precision: int = 32,
@@ -129,8 +131,10 @@ class Trainer(
         prepare_data_per_node: bool = True,
         cluster_environment: ClusterEnvironment = None,
         amp_backend: str = 'native',
-        amp_level: str = 'O2',  # backward compatible, todo: remove in v1.0.0
+        amp_level: str = 'O2',
         overfit_pct: float = None,  # backward compatible, todo: remove in v1.0.0
+        log_save_interval: Optional[int] = None,  # backward compatible, todo: remove in 0.11
+        row_log_interval: Optional[int] = None,  # backward compatible, todo: remove in 0.11
     ):
         r"""
         Customize every aspect of training via flags
@@ -178,9 +182,13 @@ class Trainer(
             distributed_backend: The distributed backend to use (dp, ddp, ddp2, ddp_spawn, ddp_cpu)
 
             early_stop_callback (:class:`pytorch_lightning.callbacks.EarlyStopping`).
-                Deprecated since v0.10.0 and will be removed in v1.0.
+                .. warning:: .. deprecated:: 0.10.0
+
+                    Will be removed in v1.0.
 
             fast_dev_run: runs 1 batch of train, test and val to find any bugs (ie: a sort of unit test).
+
+            flush_logs_every_n_steps: How often to flush logs to disk (defaults to every 100 steps).
 
             gpus: number of gpus to train on (int) or which GPUs to train on (list or str) applied per node
 
@@ -196,7 +204,12 @@ class Trainer(
 
             log_gpu_memory: None, 'min_max', 'all'. Might slow performance
 
-            log_save_interval: Writes logs to disk this often
+            log_every_n_steps: How often to log within steps (defaults to every 50 steps).
+
+            log_save_interval: How often to flush logs to disk.
+                .. warning:: .. deprecated:: 0.10.0
+
+                    Use `flush_logs_every_n_steps` instead. Will remove v0.11.0.
 
             prepare_data_per_node: If True, each LOCAL_RANK=0 will call prepare data.
                 Otherwise only NODE_RANK=0, LOCAL_RANK=0 will prepare data
@@ -235,7 +248,10 @@ class Trainer(
             resume_from_checkpoint: To resume training from a specific checkpoint pass in the path here.
                 This can be a URL.
 
-            row_log_interval: How often to add logging rows (does not write to disk)
+            row_log_interval: How often to log within steps.
+                .. warning:: .. deprecated:: 0.10.0
+
+                    Use `log_every_n_steps` instead. Will remove v0.11.0.
 
             sync_batchnorm: Synchronize batch norm layers between process groups/whole world.
 
@@ -261,6 +277,19 @@ class Trainer(
                     Defaults to `default_root_dir`.
         """
         super().__init__()
+
+        # deprecation warnings
+        if row_log_interval is not None:
+            warnings.warn("Argument `row_log_interval` is deprecated in v0.10, use `log_every_n_steps` instead."
+                          " It will be removed in v0.11.0.", DeprecationWarning)
+            log_every_n_steps = row_log_interval
+
+        if log_save_interval is not None:
+            warnings.warn(
+                "Argument `log_save_interval` is deprecated in v0.10, use `flush_logs_every_n_steps` instead."
+                " It will be removed in v0.11.0.", DeprecationWarning
+            )
+            flush_logs_every_n_steps = log_save_interval
 
         # init connectors
         self.dev_debugger = InternalDebugger(self)
@@ -299,7 +328,7 @@ class Trainer(
             process_position,
             default_root_dir,
             weights_save_path,
-            resume_from_checkpoint
+            resume_from_checkpoint,
         )
 
         # hook
@@ -310,18 +339,12 @@ class Trainer(
 
         # init data flags
         self.data_connector.on_trainer_init(
-            check_val_every_n_epoch,
-            reload_dataloaders_every_epoch,
-            prepare_data_per_node
+            check_val_every_n_epoch, reload_dataloaders_every_epoch, prepare_data_per_node
         )
 
         # init training tricks
         self.training_tricks_connector.on_trainer_init(
-            gradient_clip_val,
-            track_grad_norm,
-            accumulate_grad_batches,
-            truncated_bptt_steps,
-            terminate_on_nan
+            gradient_clip_val, track_grad_norm, accumulate_grad_batches, truncated_bptt_steps, terminate_on_nan
         )
 
         # init accelerator related flags
@@ -351,7 +374,7 @@ class Trainer(
         self.profile_connector.on_trainer_init(profiler)
 
         # init logger flags
-        self.logger_connector.on_trainer_init(logger, log_save_interval, row_log_interval)
+        self.logger_connector.on_trainer_init(logger, flush_logs_every_n_steps, log_every_n_steps)
 
         # init debugging flags
         self.debugging_connector.on_init_start(
@@ -361,7 +384,7 @@ class Trainer(
             limit_test_batches,
             val_check_interval,
             overfit_batches,
-            fast_dev_run
+            fast_dev_run,
         )
 
         # set precision
@@ -524,13 +547,15 @@ class Trainer(
                 met_min_steps = self.global_step >= self.min_steps if self.min_steps else True
 
                 if self.should_stop:
-                    if (met_min_epochs and met_min_steps):
+                    if met_min_epochs and met_min_steps:
                         self.train_loop.on_train_end()
                         return
                     else:
-                        log.info('Trainer was signaled to stop but required minimum epochs'
-                                 f' ({self.min_epochs}) or minimum steps ({self.min_steps}) has'
-                                 ' not been met. Training will continue...')
+                        log.info(
+                            'Trainer was signaled to stop but required minimum epochs'
+                            f' ({self.min_epochs}) or minimum steps ({self.min_steps}) has'
+                            ' not been met. Training will continue...'
+                        )
 
             # hook
             self.train_loop.on_train_end()